audio: platform: clean up platform driver to support future platforms
BZ: 15815

Platform driver contains Medfield specific code which needs to be removed so
that other platforms can be supported with same platform driver.

This change makes the SST driver register to the platform driver instead of the
other way round. The intel_sst_set_pll() function is replaced by the
intel_scu_ipc_set_osc_clk0() function and a proper calling mechanism is set up
in the machine driver. The intel_sst.h file is replaced by sst_platform.h. Some
ASoC core changes are backported from the upstream tree.  Clock and PLL
enabling/disabling sequence is changed.

Reviewed-on: http://android.intel.com:8080/33673
Tested-by: Hibare, PramodX <pramodx.hibare@intel.com>
Tested-by: buildbot <buildbot@intel.com>

+++ sound/soc/mid-x86/sst/intel_sst_dsp.c
@@ -38,8 +38,8 @@
 #include <linux/firmware.h>
 #include <linux/dmaengine.h>
 #include <linux/intel_mid_dma.h>
-#include <sound/intel_sst.h>
 #include <sound/intel_sst_ioctl.h>
+#include "../sst_platform.h"
 #include "intel_sst_fw_ipc.h"
 #include "intel_sst_common.h"
 #include <linux/sched.h>
@@ -241,7 +241,7 @@ static int sst_parse_module(struct fw_module_header *module,
 		}
 		/*converting from physical to virtual because
 		scattergather list works on virtual pointers*/
-		ram = phys_to_virt(ram);
+		ram = (int) phys_to_virt(ram);
 		sst_fill_sglist(ram, block, &sg_src, &sg_dst);
 		block = (void *)block + sizeof(*block) + block->size;
 	}
